pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dynamic_qrcode_image"
version = "0.0.5"
authors = [
    { name="Duke Bode", email="DukeBode@hotmail.com"},
]
description = "dynamic_qrcode_image"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python :: 3.11"
]
dependencies = [
    "flask",
    "pillow",
    "qrcode",
    "pyzbar"
]
[project.urls]
"Homepage" = "https://github.com/DukeBode/dynamic-qrcode-image"
"Bug Tracker" = "https://github.com/DukeBode/dynamic-qrcode-image/issues"
